+++ packages/dataviews/src/components/calendar/date-calendar/index.tsx
@@ -12,7 +12,7 @@ import { useCallback } from '@wordpress/element';
  */
 import { COMMON_PROPS } from '../utils/constants';
 import { clampNumberOfMonths } from '../utils/misc';
-import { useControlledValue } from '../../utils/hooks';
+import { useControlledValue } from '../utils/use-controlled-value';
 import { useLocalizationProps } from '../utils/use-localization-props';
 import type { DateCalendarProps, OnSelectHandler } from '../types';